@@ -1,7 +1,7 @@
 package com.mytlogos.enterprise.background;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class Resource<T> {
     @NonNull
