@@ -1,22 +1,16 @@
 package com.mytlogos.enterprise;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 
-import com.mytlogos.enterprise.ui.AddMedium;
-
-public class AddMediumActivity extends AppCompatActivity {
+public class AddMediumActivity extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.add_medium_activity);
+    }
 
-        if (savedInstanceState == null) {
-            getSupportFragmentManager()
-                    .beginTransaction()
-                    .replace(R.id.container, AddMedium.newInstance())
-                    .commitNow();
-        }
+    @Override
+    protected int getContent() {
+        return R.layout.add_medium_activity;
     }
 }
