@@ -55,22 +55,11 @@ public class ClientMediaList {
 
         ClientMediaList that = (ClientMediaList) o;
 
-        if (getId() != that.getId()) return false;
-        if (getMedium() != that.getMedium()) return false;
-        if (getUserUuid() != null ? !getUserUuid().equals(that.getUserUuid()) : that.getUserUuid() != null)
-            return false;
-        if (getName() != null ? !getName().equals(that.getName()) : that.getName() != null)
-            return false;
-        return Arrays.equals(getItems(), that.getItems());
+        return getId() == that.getId();
     }
 
     @Override
     public int hashCode() {
-        int result = getUserUuid() != null ? getUserUuid().hashCode() : 0;
-        result = 31 * result + getId();
-        result = 31 * result + (getName() != null ? getName().hashCode() : 0);
-        result = 31 * result + getMedium();
-        result = 31 * result + Arrays.hashCode(getItems());
-        return result;
+        return getId();
     }
 }
