@@ -1,34 +1,44 @@
 package com.mytlogos.enterprise.background.api.model;
 
+import androidx.annotation.NonNull;
+
+import java.util.Objects;
+
 public class InvalidatedData {
     private int mediaId;
     private int partId;
+    private int newsId;
     private int episodeId;
     private boolean userUuid;
-    private int externalUuid;
+    private String externalUuid;
     private int externalListId;
     private int listId;
+    private String uuid;
 
-    public InvalidatedData(int mediaId, int partId, int episodeId, boolean userUuid, int externalUuid, int externalListId, int listId) {
+    public InvalidatedData(int mediaId, int partId, int newsId, int episodeId, boolean userUuid, String externalUuid, int externalListId, int listId, String uuid) {
         this.mediaId = mediaId;
         this.partId = partId;
+        this.newsId = newsId;
         this.episodeId = episodeId;
         this.userUuid = userUuid;
         this.externalUuid = externalUuid;
         this.externalListId = externalListId;
         this.listId = listId;
+        this.uuid = uuid;
     }
 
+    @NonNull
     @Override
     public String toString() {
         return "InvalidatedData{" +
                 "mediaId=" + mediaId +
                 ", partId=" + partId +
-                ", episodeId=" + episodeId +
+                ", id=" + episodeId +
                 ", userUuid=" + userUuid +
                 ", externalUuid=" + externalUuid +
                 ", externalListId=" + externalListId +
                 ", listId=" + listId +
+                ", newsId=" + newsId +
                 '}';
     }
 
@@ -36,28 +46,29 @@ public class InvalidatedData {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-
         InvalidatedData that = (InvalidatedData) o;
-
-        if (getMediaId() != that.getMediaId()) return false;
-        if (getPartId() != that.getPartId()) return false;
-        if (getEpisodeId() != that.getEpisodeId()) return false;
-        if (isUserUuid() != that.isUserUuid()) return false;
-        if (getExternalUuid() != that.getExternalUuid()) return false;
-        if (getExternalListId() != that.getExternalListId()) return false;
-        return getListId() == that.getListId();
+        return mediaId == that.mediaId &&
+                partId == that.partId &&
+                newsId == that.newsId &&
+                episodeId == that.episodeId &&
+                userUuid == that.userUuid &&
+                externalListId == that.externalListId &&
+                listId == that.listId &&
+                Objects.equals(externalUuid, that.externalUuid) &&
+                Objects.equals(uuid, that.uuid);
     }
 
     @Override
     public int hashCode() {
-        int result = getMediaId();
-        result = 31 * result + getPartId();
-        result = 31 * result + getEpisodeId();
-        result = 31 * result + (isUserUuid() ? 1 : 0);
-        result = 31 * result + getExternalUuid();
-        result = 31 * result + getExternalListId();
-        result = 31 * result + getListId();
-        return result;
+        return Objects.hash(mediaId, partId, newsId, episodeId, userUuid, externalUuid, externalListId, listId, uuid);
+    }
+
+    public int getNewsId() {
+        return newsId;
+    }
+
+    public String getUuid() {
+        return uuid;
     }
 
     public int getMediaId() {
@@ -76,7 +87,7 @@ public class InvalidatedData {
         return userUuid;
     }
 
-    public int getExternalUuid() {
+    public String getExternalUuid() {
         return externalUuid;
     }
 
