@@ -1,9 +1,9 @@
 package com.mytlogos.enterprise.background.room;
 
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Update;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Update;
 
 public interface BaseDao<T> {
     @Update(onConflict = OnConflictStrategy.REPLACE)
