@@ -1,12 +1,15 @@
 package com.mytlogos.enterprise.background.room;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
 
+import com.mytlogos.enterprise.background.room.model.RoomExternListView;
 import com.mytlogos.enterprise.background.room.model.RoomExternalMediaList;
+import com.mytlogos.enterprise.model.ExternalMediaListSetting;
 
 import java.util.Collection;
 import java.util.List;
@@ -36,4 +39,25 @@ public interface ExternalMediaListDao extends MultiBaseDao<RoomExternalMediaList
 
     @Query("SELECT externalListId FROM RoomExternalMediaList;")
     List<Integer> loaded();
+
+    @Query("SELECT mediumId FROM ExternalListMediaJoin WHERE listId=:externalListId")
+    List<Integer> getExternalListItems(Integer externalListId);
+
+    @Query("SELECT RoomExternalMediaList.*, " +
+            "(SELECT COUNT(*) FROM ExternalListMediaJoin WHERE RoomExternalMediaList.externalListId=ExternalListMediaJoin.listId) as size " +
+            "FROM RoomExternalMediaList")
+    LiveData<List<RoomExternListView>> getExternalListViews();
+
+
+    @Query("SELECT RoomExternalMediaList.externalListId as listId,RoomExternalMediaList.uuid,url,medium,name,toDownload, " +
+            "(SELECT COUNT(*) FROM ExternalListMediaJoin WHERE RoomExternalMediaList.externalListId=ExternalListMediaJoin.listId) as size " +
+            "FROM RoomExternalMediaList " +
+            "LEFT JOIN " +
+            "(SELECT externalListId,1 as toDownload FROM RoomToDownload WHERE externalListId > 0) " +
+            "as RoomToDownload ON RoomToDownload.externalListId=RoomExternalMediaList.externalListId " +
+            "WHERE RoomExternalMediaList.externalListId=:id")
+    LiveData<ExternalMediaListSetting> getExternalListSetting(int id);
+
+    @Query("SELECT mediumId FROM ExternalListMediaJoin WHERE listId=:externalListId")
+    LiveData<List<Integer>> getLiveExternalListItems(Integer externalListId);
 }
