@@ -1,7 +1,7 @@
 package com.mytlogos.enterprise.background.room;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Query;
+import androidx.room.Dao;
+import androidx.room.Query;
 
 import com.mytlogos.enterprise.background.room.model.RoomPart;
 
