@@ -1,17 +1,17 @@
 package com.mytlogos.enterprise.ui;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.mytlogos.enterprise.R;
 import com.mytlogos.enterprise.viewmodel.AddListViewModel;
 
-public class AddList extends Fragment {
+public class AddList extends BaseFragment {
 
     private AddListViewModel mViewModel;
 
@@ -22,6 +22,8 @@ public class AddList extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
+        this.setTitle("Add List");
+
         return inflater.inflate(R.layout.add_list_fragment, container, false);
     }
 
