@@ -1,11 +1,12 @@
 package com.mytlogos.enterprise.ui;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.mytlogos.enterprise.R;
 
@@ -30,7 +31,7 @@ public class ReadHistoryRecyclerViewAdapter extends RecyclerView.Adapter<ReadHis
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.fragment_readhistory, parent, false);
+                .inflate(R.layout.readhistory, parent, false);
         return new ViewHolder(view);
     }
 
